(#2893) Switch from Should to FluentAssertions
This switches chocolatey.tests and chocolatey.tests.integration
to use FluentAssertions instead of Should. This is because Should
is no longer actively maintained, so it is time to replace it.

The Fluent Assertions Analyzers are also added in this commit.

Only the package and dependencies are switched. The code
changes be in the next commit.