(#508) Fix unit tests for new dependency handling
The underlying framework NuGet.Client has changed how it
resolves dependencies. Because of this we no longer have
information about the dependencies that will be installed
until everything has completed its run successfully.

As such, if a failure happens during the dependency handling
we only show the amount of total packages that was passed
into the executable, instead of the old handling that also included
the dependencies.
For successful runs, even the dependencies will be listed as
the amount of total packages.