(build) Switch to using MSBuildToolPath
When we updated to the latest version of the Chocolatey.Cake.Recipe
package, we switched to using VSWhere to identify where the correct
version of MSBuild is located, for this project. However, the
additional MSBuild task for creating the no7zip version of Chocolatey
hadn't been updated to use the MSBuildToolPath setting.  Initially,
this was fine, and didn't cause any problems, however, when we merged
the initial PR for the work in #508, it brought in some newer C#
syntax, which wasn't recognised by the older default version of MSBuild
which the build finds.  Switching to use MSBuildToolPath means that
everything is using the same version of MSBuild, and all builds should
completed.