(#2865) Avoid re-loading chocolatey.dll
Previously there was an issue where in certain cases, chocolatey.dll
could be re-loaded as an extension after the chocolatey licensed
extension commands are loaded, overwriting the commands already loaded.

This could be triggered by having other dlls in the extensions folder
that are not stamped with our public key, such as
chocolateygui-extension. This fix ensures that we are both not loading
specifically chocolatey.dll via our assembly resolver, and also that
extensions not stamped with our public key don't cause the main
Chocolatey assembly to be loaded as an extension dll.

ILMerged assemblies are still loaded by returning a reference to the
chocolatey assembly.