(maint) Use globbing to include test packages
This commit updates the project files for integration tests to
use a globbing pattern when determining the paths to the files
we will be using while testing the project.
This is done to make it easier to add new packages without having
to manually set each file to be copied to the bin directory.