(#2884) Add tests for config backup
In v1.2.0 choco would fall to pieces on being asked to upgrade multiple
packages in a a single command, or when upgrading a package that also
needed to upgrade a dependency. The second package installation would
fail.

This regression test ensures that we can safely `upgrade` multiple
packages at once.