(#2865) Don't resolve chocolatey for licensed ext
Previously, the change to the assembly resolver had one edge case;
if there is a chocolatey.license.xml that is found, choco will attempt
to load the licensed extension specifically. It does not check before
that whether the extension is installed or the file is anywhere it
should be expected to be, all that is offloaded to the resolver.

This change ensures that if we're looking for that extension due to
loading a valid license, we cannot return the base Chocolatey assembly
which would be completely incorrect and confuse a lot of choco's built
in license assertions and assembly version logic for that extension.