(#2930) Update GenerateDocs.ps1 for choco-theme
This file has been updated to adhere to updated style guidelines that
have been introduced in choco-theme 0.2.0. Emojis are no longer used,
and the general syntax of them has changed.

A few updates were also made to not catch `NOTE:` if it was in the
middle of a paragraph, as this would break the layout when generated.

The rules of using callouts will be available at a later date at:

design.chocolatey.org/components/callouts

The choco-theme 0.2.0 PR can be found at:

https://github.com/chocolatey/choco-theme/pull/284