(#2738) Move setup module to install .Net 4.8
This moves the setup module used during install and upgrade to
check for and install .Net 4.8 instead of .Net 4.0

The timing of the check/install is moved to before the Chocolatey
installation directory is created. This allows for re-running the
install.ps1 script if a reboot is required without needing a manual
cleanup of the partial installation.

This also adds a check for the other reboot required exit code of
1641, and throws the install/upgrade if the installer requires
a reboot.