(#2738) Check installed .NET version installed is at least 4.8
This adds a registry check to chocolatey.console to validate that .Net
4.8 or later is installed. This check is added because Chocolatey CLI
can start to execute but runs into errors when the system has an older
version of .Net framework installed. The check is very early on in the
execution so as to print out a helpful error before any errors due to
incompatibilities with the older .Net version show up.