(#1541) Add tests for future source priority feature
This commit adds new tests for when we have support in
Chocolatey CLI in either the 1.x support branch or in
the develop vNext branch.

These tests are mostly ignored for now but can be easily
changed by updating the SourcePriorityAttribute in the
tinyspec file to inherit from a CategoryAttribute to make
these run.