(#508) Fix nuspec casing on non-Windows
The NuGet.Client libraries will install nuspec files with a lowercase
filename, even if the package ID contains uppercase character(s) and
the nuspec in the nupkg is also uppercase. This behavior means that on
non-Windows platforms, the nuspec is not able to be found, because it
has incorrect casing in the filename.

This corrects that behavior on non-Windows platforms by renaming the
nuspec to the correct cased filename. On Windows, there is no need
to adjust the casing, because the filesystem is case insensitive.