(#1174)(#1610) Add tests for packing with leading zeros
These tests are to ensure that the version normalization on pack with the
new NuGet assemblies is consistent into the future. There were already tests
for removing semver v2 build metadata and that four part versions still work.
This adds tests to ensure that leading zeros are removed, and that a zero in
the fourth part of the version will be trimmed down to three parts.
https://learn.microsoft.com/en-us/nuget/concepts/package-versioning#normalized-version-numbers