(#3000) Add new rule service to handle custom rules
This commit adds a new rule service together with some base types
that will be used by the service itself, and any implementing
rule.
This service takes a path to a `nupkg` file or to a `nuspec` file
and passes along the appropriate values necessary
to validate the metadata that is being used for the
package.