(#3008) Remove support for passing in nupkg file
This commit updates the handling of installation and
upgrades of a package to first validate whether a path to a
nupkg or nuspec file is passed in or not.

If a nupkg is passed in we will throw an exception about this
and when possible give an example of how the user
may be able to install the package.
If a nuspec is passed in, we throw a different exception
about the need to package the metadata file before trying
the installation.