(#508) Update methods to work with ordering/count
There have been some upstream changes in the NuGet.Client assemblies to
provide the ability to both order by DownloadCount of packages, as well
as to provide information about the Count of packages for a given
search.

With these added, we need to taken advantage of these methods.

The commit makes use of the new SearchCountAsync method, which is then
later used in things like Chocolatey GUI.

In addition, it changes the default to use the searchResource instead
of the listResource.  When a given source doesn't have a searchResource
available, it will now fallback to the listResource, rather than the
other way around.

Finally, the paging of a search is now assumed to be provided via the
searchResource, rather than directly in Chocolatey CLI.  This change
was made given the complexities of expected results when multiple
sources are used, i.e. what constitutes a single page of results.
Additional work may be required in this area once we do more testing.