(#508) Add value to detect if page size has been specified
This commit adds a new configuration value that will be set
when a user explicitly sets the size of a page when they
search for packages on the command line.

This is done so we later can make a decision of how many
packages per source should be returned to the user.