(#508) Warn when package threshold is hit
This commit adds a warning that will be outputted to the
user when the threshold of the amount of packages
that will be searched for has been hit, or is very close to being hit.

This gives information to the user that they need to make use of paging
or refine their search to get any more results.