(#508) Update logic of how to get packages from sources
This commit updates the handling of how packages are
being queried on the specified sources, or all sources.
This is done as there are known feeds that may cause
issues if we allow any arbitrary take number to the sources.
Especially for Chocolatey Community Repository which will
have missing packages if paging is not done correctly.