(#508) Ensure that all necessary files are deleted
In version 1.x of Chocolatey, i.e. where NuGet.Core was being used,
within the remove_nuget_cache_for_package method, only one file, the
nupkg was deleted, as this was all the was required.

Starting with v2.x where the NuGet.Client assemblies are being used, it
is necessary to delete some additional files, i.e. the .nupkg.metadata
and nupkg.sha512 files.  If these files are still in place when trying
to install the package again, the nupkg is not re-downloaded as it
should be.

In addition, the download location for the nupkg file is altered to
take into consideration the Temp location as controlled by
NuGetEnvironment class, rather than the cache location.  The result
here is that all nupkgs, and related files, are collected into one
folder, rather than being located at the root of temp folder.