(#508) Add local caching of repositories
Each time the NugetCommon.GetRemoteRepositories method is called, a new
repository is created from the Factory method. This causes a problem,
since the GetRemoteRepositories method is called multiple times, which
means work is being done that is not required.  Instead, cache the
created repository in a ConcurrentDictionary, and if it exists, return
that, rather than creating a new one.

The need for this was found during testing of Chocolaty LIcensed, where
modifications are being made to the resolved repository, which are then
blown away on the next call to GetRemoteRepositories. This change means
that those changes will be persisted.