(#508) Clear repository cache when running tests
The repository caching that was introduced in previous commit causes a
problem during integration testing. This is likely due to the fact that
there are direct calls to the Service methods, rather than individual
executions of choco.exe.

This commit adds a new ClearRepositoriesCache method into NugetCommon,
which can be called in the test setup so that it is fresh for each
execution.