(#3020) Uses the Dotnet Image as the basis for the Windows image
This should fix the Docker Windows for Chocolatey 2.0+, which is currently failing due to the inability to install 4.8 during the Docker steps.

On top of this, the Arg has been misbehaving with Amazon's version of Docker. As we only intend to build the 2016 LTSC version at the moment, I have removed this until the issue is solved.