(#2953) CygwinService: always set RootDirectory
Currently, ISourceRunner instances seem to be managed as transient.
This is reasonable for many things, but it means we cannot rely on
storing state in instance fields for ISourceRunner instances at the
current time across separate invocations from the
ChocolateyPackageService.

This change ensures that the RootDirectory is always null-checked and
set if not present.

In future, we may want to discuss a design improvement/refactor to ensure that we
aren't losing any other critical state data between these invocations
for any other ISourceRunners.