(#508) Fix finding of dependency name in regex
This commit improves the regex used to extract the dependency name, by
ensuring that - and . are allowed to be included in the dependency
name.  This fixes an issue where the error output includes something
like the following:

Unable to resolve dependency ()

i.e. where the dependency name was not output.