(#508) Dependency resolution being too aggressive
There are times when the dependency resolution, and the attempt to find
a valid solution for package installation is too aggressive.  For
example, if you install a package and omit the installation of the
dependencies, then any follow up attempt to install a package will
fail, since it is checking for all dependencies to be met for all
installed packages.  This is not always valid, since it should only
worry about the package that is being installed/upgraded.

This commit adds a check to see if the package being installed has any
dependencies, and if it doesn't, don't check for dependencies for the
locally installed packages.  On top of this, check to also see if the
package being installed is an existing dependency on an existing
installed package.  If this is the case, then go through normal
dependency resolution.