(#508) Don't uninstall dependencies by default
When performing a choco uninstall operation, we don't want to, by
default, remove the dependencies for a package. This is due to the fact
that those dependent packages may be dependencies of another package.
However, we do want to be able to control when this removal of
dependencies does happen, and this is via the --force-dependencies
option. In addition, the --force option will be used to force the
removal of a dependency, even if it is a dependency of another package.

This commit makes the uninstall functionality the same as that which
exists in the current version of Chocolatey, before the uplift to use
the NuGet.Client assemblies.