(#508) Fix NRE when performing upgrade operation
When upgrading a package, there can be a problem when the dependencies
of that package has changed. For example, when upgrading the vlc package
from version 3.0.7, the package changed into a meta package, and took a
new dependency on the vlc.install package. When the upgrade is
attempted, there is a Null Reference Exception when looping through
the
resolvedPackages, since it can't find the vlc.install package in the
packagesToUninstall list, as it never existed.

This commit addresses this problem by moving the logic for uninstalling
the package to be within the packageToUninstall != null check.