(#508) Use empty array for requiredPackageIds
When resolving what packages needt to be upgraded/installed, we were
previously passing in a list of the currently installed local packages.
This was used as an attempt to ensure that these packages remained
being installed, after the operation was performed.  The side effect of
this though was that the install/upgrade operation was too aggressive,
and failed the installation/upgrade if an unrelated dependency was
missing.

For example, if you were to install Chocolatey GUI (which takes a
dependency on dotnetfx), and then you were to manually delete the
dotnetfx package, and then attempt the installation of git (which
doesn't take a dependency on dotnetfx), the install would fail saying
that there was a missing constraint. However, this is an unrelated
constraint to the package that is being installed, and therefore
shouldn't happen like this.  The current production version of
Chocolatey correctly installs git, without erroring on the lack of the
dotnetfx package.

This commit aims to address this problem by not having the currently
installed packages as required ids, and instead, leaving the
packagesConfig and targetids, to control what ends up being installed.

This change results in a new warning being shown, identifying that
there is a missing constraint for the currently installed packages, but
after discussion, it was agreed that this warning is valid, and it is
correct that it is being displayed.