(#508) Make use of new overload for ResolvePackages
In NuGet.Client, there is a new overload for ResolvePackages, which
includes whether or not prerelease packages should be included or not.
In order to make use of this, we need to pass through a
ChocolateyConfiguration object to look at whether the Prerelease option
has been set or not.

This commit addresses an issue when installing say Chocolatey GUI,
where the dotnetfx package is installed as a prerelease version, rather
than the expected stable version.