(#508) Fix versioned search using wrong queries and results
This commit fixes an issue with the search command when using `--version`
causes each query being used to always set `IsLatestVersion` or
`IsAbsoluteLatestVersion` in every call.
In way it had to be fixed, it also fixes a small bug in v1 that returns the wrong
results when a pre-release is searched for without using the `--pre` flag
and when using paging at the same time not finding the correct results.