(#508) Add version to search order by
When searching for a specific version in Chocolatey 1.x, the orderby of
the search was changed to be "Id,Version desc", and we want to use the
same orderdy when using the new NuGet.Client assemblies.

This commit switches to using the new Version property in the
NuGet.Client assemblies.

NOTE: In order for this change to work, an uplift to the NuGet.Client
assemblies will be required. A PR is open to provide this
functionality, and therefore hasn't been brought in yet.