(#1054) Ensure backups & beforeModify for dependencies
Previously we only ran backups and beforeModify scripts on the target
package(s) and not their dependencies.

This change should ensure we are _always_ running backups on any
existing packages from a NuGet source before we try to modify them in
any way. Additionally, this should ensure we can run the beforeModify
scripts for ALL dependencies of the target package(s).