(#508) Remove commented out code
This section of code had been left commented out as it dealt with the
UsePackageRepositoryOptimizations feature, which had yet to be put back
in place.  This commit removes this commented out code, as this has now
been re-implemented via the new PackageAsync method that has been added
to the ListResource within the NuGet.Client codebase.