(#508) Extend find_package to use optimizations
In order for the new Chocolatey CLI to function in the same way as the
previous 1.x versions, we need to take into account the
UsePackageRepositoryOptimizations feature.  WHen this is enabled, as
few queries as necessary should be used.  For example, when querying
for an exact package id, or checking to see if a new package version is
available, we can make a single query to a source via the Packages()
API, rather than doing multiple calls to the FindPackagesById endpoint.
 This greatly reduces the number of queries that are required overall.

This commit changes the find_package method to make use of a new
PackageAsync method which has been added to the ListResource within the
NuGet.Clint assemblies, which provides this optimization.

In addition, this commit makes a slight refactoring to the code where
find_package is called, and removes unnecessary calls to the
GetAllVersionsAsync method.  This work is exactly the same as the work
done in the GetMetaDataAsync method of the find_package method, and
therefore not required.