(maint) Use specific version for dependencies
This commit adds a specific version number in each dependency in each
of the nuspec files.  This is necessary, since moving into new
NuGet.Client assemblies, the version number is a required field in the
dependency element, and we have found issues where repository managers
don't correctly show/find the dependencies on a V3 feed, when a version
isn't available in this element.  There may be other work that needs to
be done in Chocolatey to account for this new requirement, but updating
all our internal testing packages to have a version number for each
dependency will ensure that some automated tests work as expected.