(maint) Fix logical operation to shortcircuit
This commit fixes a logical operation that had previously been
added to make use of the short circuit variant of the same logical
expression.

This is done to prevent the need to evaluate both of the statements
on each side of the or operator in the cases where pre release is already
marked as true, or when all versions is true already.