(#508) Control packages used for requiredPackageIds
When resolving what packages need to be upgraded/installed, we were
previously passing in a list of the currently installed local packages.
This was used as an attempt to ensure that these packages remained
being installed, after the operation was performed.  The side effect of
this though was that the install/upgrade operation was too aggressive,
and failed the installation/upgrade if an unrelated dependency was
missing.

For example, if you were to install Chocolatey GUI (which takes a
dependency on dotnetfx), and then you were to manually delete the
dotnetfx package, and then attempt the installation of git (which
doesn't take a dependency on dotnetfx), the install would fail saying
that there was a missing constraint. However, this is an unrelated
constraint to the package that is being installed, and therefore
shouldn't happen like this.  The current production version of
Chocolatey correctly installs git, without erroring on the lack of the
dotnetfx package.

This commit aims to address this problem by only passing the packages
to the resolver context where there are no missing dependencies. This
has fixed the issue mentioned above, and also the integration tests
which were previously failing when the last attempt was made to address
this problem:

https://github.com/chocolatey/choco/commit/6a13edc49de1c5627bf239f5a63eb37b0f297d62

Co-authored-by: AdmiringWorm <kim@chocolatey.io>