(#508) Fix invalid sequence exception on dependency not found
This commit updates the resolving of dependency to give a more accurate
information to the user when we are unable to resolve
dependency. This failure to resolve the dependency may
happen in multiple cases, but the only known case
at the moment is when a parent package is pinned and we
try to upgrade one of its dependencies.