(maint) Minor updates to tests Vagrantfile
Some of the artifacts from the Vagrantfile provisioning script
previously copied to the root of the repository. This updates to output
to the tests directory because that is where you'll be when running the
Vagrant environment, so this prevents the need to change directories.

Updates the build logic to output to a `.log` file instead of `.txt` as
`.log` is already ignored by git, and it is in fact a log. Removed the entry
from gitignore

Copies the `testInvocations.log` from the Vagrant environment into the
tests directory for analysis if desired.

Update `robocopy` call to not copy attributes as if you run a build on
macOS, this will copy the hidden attribute on any dot file and the build
will fail.