(tests) Cache package and version info
The `Test-PackageIsEqualOrHigher` and
`Test-ChocolateyVersionEqualOrHigherThan` helpers previously asked
Chocolatey for the information every time it was invoked. Rudimentary
testing (adding a `Write-Host` to `Invoke-Choco`) showed this was
happening 50+ times for some Test Kitchen runs. This commit updates the
script to store the packages in a script-level variable that is
referenced in future calls to reduce the number of times we call
`choco.exe`.

The variables used are purposely long to avoid collisions with any
testing variables.