(#508) Fix repository optimizations for NuGet V3
When we re-introduced repository optimizations for Chocolatey, there
was an assumption that there would always be a ListResource available
for the feed.  However, this has proven not to be the case.  For
example, when using something like Sleet or Baget, which are V3 only
feeds, the ListResource does not exist.  In these cases, we should
defer to using the non-optimized queries.ThresholdHit

This commit ensures that this is done, by only using the optimized
query, when it is known that there is a ListResource available.