(#2642) Remove deprecated shims
The shims for chocolatey, cinst, clist, cpush, cuninst, and cup, where
deprecated in a previous release of Chocolatey, and this commit removes
them, so that they will not appear in the upcoming 2.0.0 release.

Also the documentation around the deprecations have been removed, and
tests updated to reflect that these shims are no longer included.