(#158) Remove list alias for remote searches
(#158) Remove list alias for remote searches

This commit removes the list alias for remote searches
and renames the command class to be ChocolateySearchCommand
to reflect it is for remote searches.

As a side effect, the Info command now inherits from
the search command which should be fine in most
cases.