(#2902)(#2585) Update TLS protocol handling
This updates the TLS protocols that Chocolatey sets now that Chocolatey
is built with .NET 4.8.

First, this removes the fallback and warning about being unable to set
TLS 1.2, as .NET 4.8 will always have TLS 1.2 available. Second, it
uses the TLS 1.2 definition directly available in the system libraries.
Third, it removes the provideWarning bool in the method definition, as
the warning is no longer applicable.

Finally, a OS version check is added, so the system default will be
used on Windows 8/Server 2012 and newer, as per Microsoft guidance,
but explicitly enabling TLS 1.2 on older operating systems. See code
comments for links to documentation.