(#3088) Don't loop over null hook script variables
PowerShell 2 had the behaviour that foreach would iterate over a null
variable
(https://devblogs.microsoft.com/powershell/new-v3-language-features/).
This has resulted in an issue if no hook scripts are passed in, then
chocolateyScriptRunner would cause an error. This update changes this to
do a nullcheck like we already had for the packageScript parameter.