(#2716) Remove storing arguments and vars w/spaces
We need to ensure we're not storing sensitive arguments during tests.
Until we have a reasonable way to filter those out, not going to store
the arguments in the invocation logs. Remove spaces from the caching
variable names to avoid mistakes as spaces in variable names are
uncommon in PowerShell.