(#3097) Remove ancient history from the codebase
A lot of our builtin documentation has mention of very old versions of
Choco when features were introduced. Given we no longer support those
versions, all supported versions contain the relevant features, so it's
superfluous for us to mention the exact versions.

Removing these removes a lot of clutter when looking for command-line
help and should also cleanup our generated online documentation as well.
No user on a current version of chocolatey is likely to care what was
introduced in v0.x of Chocolatey; it's available in the current version,
and that's ultimately all that matters.

For our docs website it makes more sense to care, but still not enough
for us to worry about telling folks that something was introduced before
v1, much less a specific 0.x version.