(maint) Add chocolatey license to gitignore file
There are some scenarios where a developer may have a Chocolatey License
in the working directory (for instance testing the output of Chocolatey
when a license is present but the extension isn't, or perhaps adding a
new license command). This adds license files to gitignore so you don't
accidentally commit one to the repo.