(maint) Fix null config value throwing exception
When one of the custom classes used as in ChocolateyConfiguration
is set as null, the code we have will still try to recursively call
and set any parameters on this null class.

This will cause an exception to happen as the object used is null
and can not be used when setting a value.