(#2345) Avoid any trailing spaces after silentArgs
Fix a bug where concatenating silentArgs and additionalInstallArgs adds
a trailing space at the end of silentArgs even when
additionalInstallArgs is empty. Such trailing space can change the
meaning of the last positional argument (that may contain spaces).

Final arguments are now created as space joined string arrays filtered
down to only non-empty items. Fix is applied to all installer types.

It is expected not to manipulate the arguments themselves in case there
are any spaces that the package maintainer has included.