(tests) Add test for missing lib directory
In CLE, there are cases where having a missing lib directory will break
some commands.

However, the behaviour still needs to be asserted for CLI to
ensure that we have the correct test here,
and we will also run these tests under CLE
in order to ensure that adding the extension does not break this edge
case anymore.