(#2815) Add ability to get specific feature status
In addition to being able to get the complete list of features on the
system, it should be possible to get the status of a specific feature.
This matches what is possible for example with the choco config get
command.

This commit adds the ability to use choco feature get to get the
current Enabled/Disabled status for a given feature.  If the feature
name provided doesn't exist an error will be returned.

Co-authored-by: Tom Vervoort <tomv@infosupport.com>