(#158) Add handling for -i option
When the work was done to change the behaviour of choco list to only
include information about local packages, we missed the ability to run
a command like choco list -li.  Historically, this command would have
listed all locally installed packages, as well as information about
applications that have been directly installed outside of Chocolatey.
This was due to the fact that since -l is no longer a parsed argument.
the argument parser doesn't interpret the other options.

This commit addresses this by adding specific handling for the option
combinations that were previously documented on our site, showing a
warning when they are used, and directly setting the
IncludeRegistryPrograms configuration value when required.

This code will be removed in a future version of Chocolatey, most
likely in V3.