(#2219) Move named source logic to Validate method
Having this logic in the ParseAdditionalArguments method makes the
command useless if there isn't a --source or defaultPushSource set.
The error that would occur happens before the command can even output
help information, and this also interferes with licensed functionality.

This change moves the checks into the Validate method and consolidates
the validation slightly as there was some double-work being done with
the validation logic being somewhat split between the two methods.