(tests) Adjust choco help tests
Avoid tests calling `choco --help --help` etc unnecessarily.

Don't copy config/packages before running help tests, we don't want any
non-default config interfering in the help
tests.
They shouldn't, but these tests were expected to fail with the issue
that the prior commits resolve, and they didn't.

So it seems like these tests are making the assumption that config
doesn't matter for their operation, and it shouldn't. But we should test
on a default config just to be sure, and catch any issues like this in
future.

Also fixed a broken regex that was making these tests not run at all
thanks to our trimming output lines with Invoke-Choco. Oops!