(#1720) Parse System Proxy setting for protocols
The System Proxy setting could be set for different protocols. When that
happens it is listed something like:
`http=server:port;https=server:port;ftp=server:port;socks=server:port`.
If the string contains a semicolon (`;`), we should split on that and
parse out the https option if it exists. If it doesn't exist we should
parse the http, leaving no proxy if none is set for either http or
https.