(#63) Revert push --source alias changes
We have decided that the current implementation doesn't work as we would
want for some common use cases, and so this commit removes the changes.

Tests have been left behind and flagged as ignored, so we can make use
of them when a design is agreed on and the functionality is
reimplemented.