(#1720) Handle multi-protocol proxy configurations
It is possible to configure protocol specific proxy settings. When you
do that, you could conceivably set the protocol for ftp only. We need to
ensure we're only pulling the http or https proxy. You could also set it
just for http, in which case we would be taking `http=Something` and
applying it when we should only be applying `Something`. This updates to
handle the scenarios where that ends up in the configuration.