(#508) Set OrderBy property for count query
When using the SearchCountAsync from the NuGet.Client codebase, we want
the query that is used to get the count to be as close to the outgoing
search query as possible.  As such, make sure to set the OrderBy
property.

This commit ensures that the outgoing query matches the one that is
used in Chocolatey CLI 1.x.