(#508) Update to latest NuGet.Client packages
This includes a change to how the count query is generated, so that we
can be sure that we are emitting the same query as was done in v1.x.
A problem with this query was noticed when testing Chocolatey GUI, and
pages of packages weren't able to be navigated, since the total number
of packages was not being returned correctly from the server.