(build) Switch to using Chocolatey specific build
There is now a much reduced version of the build for NuGet.Client,
which only builds the components that are requried for Chocolatey. This
was necessary since the build started failing both internally and on
GitHub Actions, when attempting to build the VIsual Studio extension
and associated projects.  Since these projects are not needed for the
Chocolatey integration, they have been removed from the build.  The end
result is that the build is much quicker, and builds both locally and
on CI.