(#2690) Add backup/restore strategy method
This commit updates the files service to introduce a new method
that will handle the logic of making backups, or restore said backups
when there is a need.

This is done to prevent the need to update the logic in multiple places, and
to make the backup/restore compatible with what we are documenting.