(#2690) Fix backup path used for transform service
The transform service was using the old location where we previously
stored backed up packages.

This new logic does its best attempt of acquiring the new location of the
backup folder, while ignoring any backed up folders of the package that is
being installed.

This may have the side affect that forced installations will not use
a previously created XDT transformation, but only new files in the package.