(tests) Ensure tab completion doesn't emit errors
There are a couple of expected errors for tab completion to hit, that
we are handling but don't remove from $error when handling those.

This is actually causing a few tests to fail, and in general it's
best not to pollute $error with already-handled errors.
Since we can't use -ErrorAction Ignore for PSv2 compat reasons, ensure
we catch those errors instead of continuing and just remove them from
the $error variable before continuing.