(tests) Remove broken test that was never working
This test was so broken it wasn't even failing, and attempting to fix it
made things worse. After speaking to Kim, we're best off removing it for
now as it's not really testing what we'd want it to do and if we want a
test for this in future we'll need to write it from scratch anyway.