(#508) Add try catch around uninstall resolver
THe uninstall resolver has the potential to throw exceptions when
it finds a package that can not be uninstalled due to it being a
dependency on other packages, as such we need to be able to
handle these scenarios and not stop further processing of the
additional package names.

The changes in this commit adds a try catch block around the resolver,
and if it fails a new package result is created with the reason of the
failure.